+++ src/main/kotlin/site/siredvin/peripheralium/data/ModDataGenerationEntrypoint.kt
@@ -6,5 +6,6 @@ import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator
 class ModDataGenerationEntrypoint: DataGeneratorEntrypoint {
     override fun onInitializeDataGenerator(fabricDataGenerator: FabricDataGenerator) {
         fabricDataGenerator.addProvider(ModRecipeProvider(fabricDataGenerator))
+        fabricDataGenerator.addProvider(ModBlockLootTableProvider(fabricDataGenerator))
     }
 }